Reorganizes fitting tracker module structure
