+++ src/easydiffraction/analysis/fitting/progress_tracker.py
