+++ tests/unit/analysis/minimizers/test_minimizer_base.py
@@ -90,7 +90,7 @@ def test_finalize_fit(mock_minimizer, mock_parameters):
     assert result.parameters == mock_parameters
 
 
-@patch('easydiffraction.analysis.minimizers.fitting_progress_tracker.FittingProgressTracker')
+@patch('easydiffraction.analysis.fitting.progress_tracker.FittingProgressTracker')
 def test_fit(mock_tracker, mock_minimizer, mock_parameters, mock_objective_function):
     mock_minimizer.tracker.finish_tracking = MagicMock()
     result = mock_minimizer.fit(mock_parameters, mock_objective_function)
